REFACTOR : Navbar #6 move Icon style from separate file to globalStyle
